Add "Dublin Circus Festival 2016" postmortem
At the time this was the biggest event I'd ever been involved in, and
although I have wanted to write about it for a very long time, following
my departure from the circus community I felt I wouldn't be able to give
it as unbiased a go as possible.

The phrasing of a lot of this article will be improved in the future.