Update DCF Postmortem, add Diabolo file
I only did a quick pass when originally releasing the Dublin Circus
Festival 2016 postmortem post, so I went through and re-phrased or cut
down the text to be less redundant overall. The wording of certain
sentences also made it feel incongruent with the overall tone.

I haven't actually written the diabolo article yet as it will likely
feature video content, but the file which will later be served now
exists as a draft with its appropriate URL slug.