Clarify ambiguity in CoC Anatomy post
It wasn't really clear in the text when I took a break; although for the
purposes of the overall piece it didn't actually matter, it was enough
to break the reading flow of at least 1 person and needed to be fixed.

I also murdered an unnecessary word in the same paragraph.