Add navbar to main container, adjust remove link icons from anchor tag
I disliked that the text-decoration was being triggered with the empty
space between the icon and the text - though it also begs the question
if the underline is necessary or not, and what alternative might be
aesthetically pleasing while still acting as a useful signifier.

Next up to the chopping block will be fonts as a whole - I need to think
about the colours and sizes a bit.