Amend password managers post with clarified intent around pricing
I very recently had the privilege of having someone review my writing,
and they pointed out that my prior phrasing of the appeal of OSS could
be taken the wrong way, particularly by those who work primarily in it.

As a result of this feedback I realised it didn't reflect my opinions
properly, and even came off dismissive, which was not my intent.

Although part of me wanted to let it stand for historic purposes, I
realised that historic accountability would always exist in the
websites' commit history, and that if it were the first written piece of
mine someone read, I would never get a chance to make a second
impression.