Change default archtype taxonomy and hide blog overview
While working on the draft for a new post I realised there was still
some redundant references to 'tags' instead of 'topics', but that I also
hadn't yet done any of the formatting to make the blog overview readable
with tags underneath. This fixes up the references, and also disables
the entire block related to topics for the time being.