Update homepage to use markdown page
This simple change makes it so that the homepage of the website is now
built from _index.md in the base content/ folder, which will make
updating the text significantly easier.

I have also added a .gitignore file so as to avoid commiting the public/
build folder, though I have a sneaking suspicion this could cause some
problems with how Netlify builds it.