Add switcher class, rename layout file
This adds the basic switcher class to convert a three column layout to a
singular column based on a threshold value through the use of
flex-basis. It's a very common layout concern I will use multiple times
on my website.

style.css has also been renamed layout.css to reflect the purpose of the
stylesheet, as although there are some layout elements left for
consideration, I would like to work on the aesthetics of the website
sooner than later, starting with the typography and colour scheme.