Rework project structure into Eleventy format
This commit adds all of the initial, small changes necessary for the
pivot into Eleventy. There's still a lot of work to go until I'm happy
with the base design of the website before I commit fully to refactoring
things into templates and includes, but this is a good start.

Once I have verified things work as expected, I will move into making
the navigation layout look and act as expected - then do some work on
the typography and overall style guide.