Add style.css, adjust navigation layout and aesthetic
I'm trying to decouple layout and style concerns in CSS, which has lead
me to creating a new file as part of this commit. Thus far little has
changed; I'm not entirely sold on the navigation layout on the basis
that didn't intend for it to be so evenly spaced.

It might merely feel weird because I haven't done it before, and also because
there's so few menu items - but in the "mobile" layout for the switcher
breakpoint, it also seem as though I should remove or reduce the margin
between items, as the spacing feels somewhat off.

Typography will be the next change, then I need to consider the colour
palette - otherwise I will actively avoid it forever.