Add content from previous iteration of website
This is most of the content from my old website - the only thing I've
opted not to include was my review of "Inclusive Components", as in
retrospect it is not particularly well-written and also doesn't fit the
content of the rest of my website.

Nothing has been changed in the existing content; I should be able to
re-use much of the frontmatter with Eleventy, but we'll see what
happens.