Restructure base layout and add collections for homepage display
These changes are designed so that in the near future it will be
possible to generate the home page content through the collections of
content by using data in their frontmatter.

Still plenty of work to do as far as structuring the actual templating
goes, but I'm happy with the progress so far and are looking forward to
the emotional catharsis when it all comes together.