Add writing and project collections to homepage
This commit creates the necessary calls in the featured partials to
filter for and grab data for the front page of the website. With these
changes, I am incredibly close to being "finished" with rebuilding the
website to where it was - with discoverable content and links.

I don't think it's clear that the titles are clickable, however, and I
may change to a more explicit colour for heading links, or a "see more"
anchor at the bottom instead.