Add basic writing page layout
This simplistic layout may end up being identical to what's used for
projects - and thus warrant a refactor, but for the time being this
works fine. The commit adds the basic layout for writing, as well as a
library to format the date appropriately as a filter for the date
derived from the page frontmatter.

There are a handful of other small things that need to be done for the
page; for one, I think a smaller maximum width would be better.
Secondly, there is a 'back to top' link missing, though that would be in
the scope of the base template. I think including a "Last updated" entry
at the bottom of the page would be neat, albeit unnecessary.