Change page layout: add directory data, change semantic element
Directory data files allow us to define things to add to the frontmatter
of all files included in a folder; thus we can use it to define a
default layout for written pieces. This change reflects that, and
removes the now-redundant layout link from the individual files.

This can be overwritten with an explicit layout from a piece. This also
changes the semantic element containing all of the content from section
to article; although I think of them more like memoirs, that is not a
useful distinction for the browser!