Add writing overview page
I'm shocked at how much of this actually worked out the gate. This
commit refactors the written pages a little bit to reflect the mental
model I have for the website: the writing page has a list of writing on
it, which are all singular articles.

Oddly, the page layout for the overview was broken until I changed the
ordered list to have list-style: none. I have no idea why; it's probably
something to do with how list styles render.

It's all beyond me; what matter is it works.