Update pages, hide project section, update writing layout
There's a huge number of changes in this commit, but many of them are
relatively small. It all revolves around "cleaning up" the website to a
presentable state (No more visible half-broken functionality); there is
still plenty more to do, but I have a lot of writing I want to get done,
so the website needs to be accessible to the public.