content: Improve phrasing of DCF Postmortem & Pobail Retro.
There was a few minor grammar issues in the Pobail Stomp retrospective
that bothered me: I originally intended to fix a blatant mistake, but I
could not find it anywhere.

I'm still unsure how I feel about editing old content: the DCF
postmortem is one of the oldest works I have kept online. I found that I
referred back to it often, so it felt appropriate to fix it.

It stands as an evergreen measure of how I have changed as an
event organiser, but now the grammar makes me cringe less.