feat: Rename rss links to feed, add placeholder xsl file
This commit shifts the atom feed back to the "feed" naming convention
instead of rss. This is futureproofing: naming it RSS was incorrect, but
"feed" is generic enough that should I explicitly choose to have it as
RSS in the future, it doesn't need to be changed.

It also adds a placeholder for an XSL file, which is broadly an XML
stylesheet. Though not strictly necessary, this is useful functionality,
so I want to implement this properly in the near future.