feat: Add global text-wrap: balance to CSS, change body to system-ui
This looks a little weird in some places, but I'm still going to set it.
text-wrap: balance attempts to balance pargraphs of text, and I think it
looks great. There's some clear improvements to be made, but I believe
they have more to do with the general typography.

I've also changed the body text to use system-ui instead of a specific
typeface for the sake of readability and accessibility. A website
shouldn't override a user's browser overrides: system-ui takes the
default interface font on a given system.

I intend to do some farther investigation to figure out precisely how
the browser detects the default interface typefaces.